Code with no bug
Take a look at the manager and see if you can undertsand what I did. I will be more than happy to show anyone who wants to know the proccess. The bug that caused settings to throw a nullpointer is now gone. Essientially each pane clears the previous and adds a new one.