New Current
Added a folder called SE 300.
What it does is when the program starts, a UserID.txt file is created and when the user successfully logins in or makes an account a copy of the userID is exported to the text file.

UserID class has methods that call for the userID which will require a try throw catch. For the catch use the alert method. The alert message should never pop up unless that temporary file containg the userID (UserID.txt) has been removed or deleted from its original location. As a result to solve that the user must restart the application. I hope it works on all your computers.